Extract renderWithRedux
