+++ src/pages/__tests__/HomePage.spec.tsx
@@ -1,14 +1,10 @@
-import 'core-js/stable';
 import React from 'react';
-import { shallow, configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 import HomePage from '../HomePage';
-
-configure({ adapter: new Adapter() });
+import renderWithRedux from '../../__tests__/renderWithRedux.spec';
 
 describe('HomePage component', () => {
   test('should render', () => {
-    const component = shallow(<HomePage />);
-    expect(component).toMatchSnapshot();
+    const { asFragment } = renderWithRedux(<HomePage />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });
